+++ packages/react/src/useYorkieDoc.ts
@@ -23,7 +23,9 @@ import {
 import pkg from '../package.json';
 import { useYorkieClient } from './YorkieProvider';
 import { useYorkieDocument } from './DocumentProvider';
-import { useMemo } from 'react';
+import { useMemo, useRef } from 'react';
+import { createDocumentStore } from './createDocumentStore';
+import { useSelector } from './useSelector';
 
 /**
  * `useYorkieDoc` is a custom hook that initializes a Yorkie Client and a
@@ -44,6 +46,24 @@ export function useYorkieDoc<R, P extends Indexable = Indexable>(
   loading: boolean;
   error: Error | undefined;
 } {
+  const documentStoreRef = useRef<
+    ReturnType<typeof createDocumentStore<R, P>> | undefined
+  >(undefined);
+
+  if (!documentStoreRef.current) {
+    documentStoreRef.current = createDocumentStore<R, P>({
+      doc: undefined,
+      root: opts?.initialRoot ?? ({} as R),
+      presences: [] as Array<{ clientID: string; presence: P }>,
+      connection: StreamConnectionStatus.Disconnected,
+      update: () => {},
+      loading: true,
+      error: undefined,
+    });
+  }
+
+  const documentStore = documentStoreRef.current;
+
   // NOTE(hackerwins): useMemo is used to prevent creating a new client
   // every time the component re-renders. If the apiKey or docKey
   // changes, a new client will be created.
@@ -53,7 +73,7 @@ export function useYorkieDoc<R, P extends Indexable = Indexable>(
       userAgent: pkg.name + '/' + pkg.version,
       ...opts,
     };
-  }, [apiKey]);
+  }, [apiKey, opts]);
 
   const {
     client,
@@ -61,24 +81,24 @@ export function useYorkieDoc<R, P extends Indexable = Indexable>(
     error: clientError,
   } = useYorkieClient(clientOpts);
 
-  const { root, presences, connection, update, loading, error } =
-    useYorkieDocument(
-      client,
-      clientLoading,
-      clientError,
-      docKey,
-      opts?.initialRoot,
-      opts?.initialPresence as P,
-    );
+  useYorkieDocument<R, P>(
+    client,
+    clientLoading,
+    clientError,
+    docKey,
+    opts?.initialRoot ?? ({} as R),
+    opts?.initialPresence ?? ({} as P),
+    documentStore,
+  );
+
+  const documentState = useSelector(documentStore);
 
   return {
-    root: root as R,
-    presences,
-    connection,
-    update: update as (
-      callback: (root: R, presence: Presence<P>) => void,
-    ) => void,
-    loading,
-    error,
+    root: documentState.root,
+    presences: documentState.presences,
+    connection: documentState.connection,
+    update: documentState.update,
+    loading: documentState.loading,
+    error: documentState.error,
   };
 }