+++ CHANGELOG.md
@@ -7,6 +7,48 @@ and Yorkie JS SDK adheres to [Semantic Versioning](https://semver.org/spec/v2.0.
 
 ## [Unreleased]
 
+## [0.6.23] - 2025-08-01
+
+### Fixed
+
+- Cleanup ESLint packages/react by @jaesoekjjang in https://github.com/yorkie-team/yorkie-js-sdk/pull/1027
+- Remove deprecated SelectOperation by @hackerwins in https://github.com/yorkie-team/yorkie-js-sdk/pull/1038
+- Resolve convergence issues in Array.Move and Array.Set by @emplam27 in https://github.com/yorkie-team/yorkie-js-sdk/pull/1037
+
+## [0.6.22] - 2025-07-28
+
+### Changed
+
+- Change VersionVector actorID encoding from hex to base64 by @mnnseong in https://github.com/yorkie-team/yorkie-js-sdk/pull/1030
+
+### Fixed
+
+- Fix eslint cleanup SDK by @pengooseDev https://github.com/yorkie-team/yorkie-js-sdk/pull/1023
+
+## [0.6.21] - 2025-07-25
+
+### Added
+
+- Add nextjs-quill example using Yorkie and Quill v2 by @miinhho in https://github.com/yorkie-team/yorkie-js-sdk/pull/1032
+- Support for object, array and enum in ruleset builder by @sigmaith in https://github.com/yorkie-team/yorkie-js-sdk/pull/1025
+
+### Changed
+
+- Migrate vanilla-quill example dependency to Quill v2 by @miinhho in https://github.com/yorkie-team/yorkie-js-sdk/pull/1034
+
+## [0.6.20] - 2025-07-17
+
+### Added
+
+- Support Undo/Redo for array.add and array.remove by @KMSstudio in https://github.com/yorkie-team/yorkie-js-sdk/pull/1014
+
+### Fixed
+
+- Update CONTRIBUTING.md to use `pnpm i` by @hackerwins in https://github.com/yorkie-team/yorkie-js-sdk/pull/1012
+- Update deprecated pprof flag by @pengooseDev in https://github.com/yorkie-team/yorkie-js-sdk/pull/1020
+- Change schema tests to use assert.include by @hackerwins in https://github.com/yorkie-team/yorkie-js-sdk/pull/1024
+- Fix typo in MaxDelimiter constant name by @jaesoekjjang in https://github.com/yorkie-team/yorkie-js-sdk/pull/1022
+
 ## [0.6.18] - 2025-07-01
 
 ### Added